feature dump: GUI / minimize to tray
