+++ .github/workflows/pyinstaller.yml
@@ -22,7 +22,7 @@ jobs:
       run: python -m pip install -r requirements.txt pyinstaller
       
     - name: Build with pyinstaller
-      run: pyinstaller main.py --onefile --hide_console hide-late
+      run: pyinstaller main.py --onefile --add-data=assets:assets --hide-console hide-late
 
     - uses: actions/upload-artifact@v2
       with:
@@ -48,7 +48,7 @@ jobs:
       run: python -m pip install -r requirements.txt pyinstaller
       
     - name: Build with pyinstaller
-      run: pyinstaller main.py --onefile --uac-admin
+      run: pyinstaller main.py --onefile --add-data=assets:assets --hide-console hide-late
 
     - uses: actions/upload-artifact@v2
       with:
